更新 config.rs

+++ AppOpt/src/config.rs
@@ -370,7 +370,7 @@ fn reinstall_inotify_watch(inotify_fd: i32) -> bool {
     };
     let inotify_wd = INOTIFY_WD.load(Ordering::Acquire);
     unsafe {
-        libc::inotify_rm_watch(inotify_fd, inotify_wd);
+        libc::inotify_rm_watch(inotify_fd, inotify_wd as u32);
     }
     let cfg_cstr = match CString::new(cfg.config_file.as_str()) {
         Ok(c) => c,
